refactor: Change code arch
