+++ configs/config.yaml
@@ -1,0 +1,3 @@
+env: "local" # local, dev, prod
+db_file: "database.json"
+source_url: "https://xkcd.com"
